backend/requirements.txt
fastapi[standard]
uvicorn
google-generativeai
firebase-admin
pymupdf
langchain-text-splitters
python-dotenv
numpy
